adding authentication logic and realted files
