+++ .gitignore
@@ -2,3 +2,6 @@ application_code/.idea/
 application_code/node_modules/
 application_code/package-lock.json
 application_code/.nyc_output/
+.DS_Store
+application_code/.DS_store
+.DS_Store
