+++ application_code/views/register.hbs
@@ -5,7 +5,7 @@
         <i class="fas fa-user-plus"></i> Register
       </h1>
 
-      <form action="/users/register" method="POST">
+      <form action="/register_user" method="post">
         <div class="form-group">
           <label for="name">First Name</label>
           <input
@@ -80,7 +80,7 @@
           Register
         </button>
       </form>
-      <p class="lead mt-4">Have An Account? <a href="/users/login">Login</a></p>
+      <p class="lead mt-4">Have An Account? <a href="/login">Login</a></p>
     </div>
   </div>
 </div>
