+++ application_code/views/welcome.hbs
@@ -5,8 +5,8 @@
       <div class="col-md-6 m-auto">
         <div class="card card-body text-center">
           <p>Create an account or login</p>
-          <a href="/users/register" class="btn btn-primary btn-block mb-2">Register</a>
-          <a href="/users/login" class="btn btn-secondary btn-block">Login</a>
+          <a href="/register" class="btn btn-primary btn-block mb-2">Register</a>
+          <a href="/login" class="btn btn-secondary btn-block">Login</a>
         </div>
       </div>
     </div>
